pyproject.toml
[project]
name = "tai-mcp-memory"
version = "0.1.0"
description = "MCP Memory Server with persistent, semantic memory for LLM agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.23.0",
    "python-dotenv~=1.2.1",
    "chromadb~=1.4.1",
    "pydantic~=2.12.5",
    "openai~=2.16.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = [".venv", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
# E402: module-level import not at top — intentional sys.path setup pattern in this project
# E501: line too long — not enforced, handled by editor/formatter preference
ignore = ["E402", "E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
explicit_package_bases = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "WARNING"

[tool.coverage.report]
fail_under = 70
omit = ["memory_server.py", "restore_memory.py", "tests/*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
